Database: Flatten field definition
